+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ### 5.1.0 - 2023/10/30 10:55
-* Feat:
+* Feat: 
   - [PAY22-2045] Magento - Mejoras para órdenes perdidas. @fcarrero [#46](https://github.com/conekta/customer-magento-plugin/pull/46)
   - Upgrade php version requerida.
   - Fix conekta.png route location
+  - cancel orders
+  - change status from suspect fraud to pending payment
 ### 5.0.5 - 2023/10/09 10:55
 * Feat:
   - [PAY22-2045] Magento - remove old sdk-php code. @fcarrero [#45](https://github.com/conekta/customer-magento-plugin/pull/45)